+++ src/main/java/org/example/Main.java
@@ -1,7 +1,5 @@
 package org.example;
 
-import java.util.Scanner;
-
 public class Main {
 
   public static void main(String[] args) {
@@ -15,73 +13,8 @@ public class Main {
     //一番やりたいこと -> 検索機能の実装
     //Libraryクラスは本棚、検索機能、検索結果の表示を行うメソッドを実装
 
-    Library library = new Library();
-    Scanner scanner = new Scanner(System.in);
+    Message.startMessage();
+    InputDataAndOutPutData.inputDataAndOutPutData();
 
-    System.out.println("書籍検索システムへようこそ！");
-    System.out.println("モードを選択してください！");
-
-    makeLine();
-    System.out.println("1：タイトルで検索");
-    System.out.println("2：著者名で検索");
-    System.out.println("3：管理番号で検索");
-    System.out.println("4：登録されている本を一覧で表示");
-    makeLine();
-
-    System.out.print("数字で選択してください : ");
-
-    String selectMode = scanner.next();
-
-    //入力値バリデーション
-    boolean isNumber = selectModeCheck(selectMode);
-
-    if (isNumber) {
-      makeLine();
-      switch (Integer.parseInt(selectMode)) {
-        case 1 -> {
-          System.out.println("タイトルで検索します。");
-          System.out.print("タイトルを入力してください。 : ");
-          String title = scanner.next();
-          library.searchBookTitle(title);
-        }
-        case 2 -> {
-          System.out.println("著者名で検索します。");
-          System.out.print("著者名を入力してください。 : ");
-          String author = scanner.next();
-          library.searchBookAuthor(author);
-        }
-        case 3 -> {
-          System.out.println("管理番号で検索します。");
-          System.out.print("管理番号を入力してください。 : ");
-          int number = scanner.nextInt();
-          library.searchBookNumber(number);
-        }
-        case 4 -> {
-          System.out.println("登録されている本を一覧表示します。");
-          library.showBooks();
-        }
-        default -> System.out.println("1〜4で選択してください。");
-      }
-    } else {
-      System.out.println("正しく入力されていません。");
-    }
   }
-
-  //数字かどうかチェックする
-  public static boolean selectModeCheck(String number) {
-    try {
-      Integer.parseInt(number); //文字列を整数に変換して例外が起こればcatchへ
-      return true;
-    } catch (NumberFormatException e) {
-      return false;
-    }
-  }
-
-  //見やすいように行の区切りを生成
-  public static void makeLine() {
-    System.out.println(
-        "**********************************************************************************************************");
-  }
-
-
 }
